Moved metric class to own file.
