+++ graph.py
@@ -4,6 +4,8 @@ import warnings  # To suppress UserWarning associated w/ using _nolegend_
 
 import pandas as pd
 
+from metric import Metric
+
 
 class Graph(object):
     """
@@ -13,57 +15,6 @@ class Graph(object):
     Functions:
 
     """
-
-    # Maybe use built in pandas/numpy data structure instead
-    class Metric:
-        def __init__(self, x_axis_label, y_axis_label, name, color, func):
-            self.x_axis_label = x_axis_label
-            self.y_axis_label = y_axis_label
-
-            # Init
-            self.__name = name  # Name of metric
-
-            self.__legend = "{}{}".format(name[0:1].upper(), name[1:])
-
-            self.__color = color
-
-            self.__func = func  # Function to generate values
-
-            # Change to map?
-            # Create cache
-            self.__cache = pd.DataFrame({self.x_axis_label: [], self.y_axis_label: []})  # Stores generated values
-
-        def get_name(self):
-            return self.__name
-
-        def get_legend(self):
-            return self.__legend
-
-        def get_color(self):
-            return self.__color
-
-        def get_func(self):
-            return self.__func
-
-        def get_cache(self):
-            return self.__cache
-
-        def set_legend(self):
-            self.__legend = "_nolegend_"
-        
-        def set_func(self, func):
-            self.__func = func
-
-        def set_cache(self, data):
-            self.__cache = self.__cache.append(data, ignore_index=True)
-
-        def generate_values(self, input_values):
-            # Turn x, y, target values into generator to access differently
-
-            generated_values = [self.__func(value) for value in input_values[self.y_axis_label]]
-
-            # Add generated data to cache
-            self.set_cache(pd.DataFrame({self.x_axis_label: input_values[self.x_axis_label], self.y_axis_label: generated_values}))
 
     def __init__(self, figure, title, x_axis_label, y_label, row, col, total_num, target=None, pan=300):
         # Configuration settings
@@ -94,8 +45,8 @@ class Graph(object):
 
         # Desired metrics
         self.metrics = {
-            'main': self.Metric(self.x_axis_label, self.y_axis_label, 'main', 'blue', lambda x: x),
-            "target": self.Metric(self.x_axis_label, self.y_axis_label, "target", "orange", lambda x: self.target)
+            'main': Metric(self.x_axis_label, self.y_axis_label, 'main', 'blue', lambda x: x),
+            "target": Metric(self.x_axis_label, self.y_axis_label, "target", "orange", lambda x: self.target)
         }
 
     def update(self, new_data):
@@ -167,26 +118,6 @@ class Graph(object):
 
         self.metrics['target'].set_func(lambda x: self.target)
 
-        # # # Cannot purge in future
-
-        # Purge old target line ?
-        #for line in self.axis.lines:
-         #   if line.get_color() == self.config['target'][0]:
-          #      self.axis.lines.remove(line)
-
-        # TODO
-        # print(self.data[self.y_axis_label].iloc[-1] if len(self.data[self.y_axis_label]) > 0 else 0)
-
-        #self.plot_line("target", pd.DataFrame(
-         #   {
-          #      self.x_axis_label: [
-           #         # FIX
-            #        self.data[self.y_axis_label].iloc[-1] if len(self.data[self.y_axis_label]) > 0 else 0,
-             #       20000],
-              #  self.y_axis_label: [self.target, self.target]
-            #}
-        #))
-
     def add_metric(self, title, func):
         """
         Use: To add new metric to graph
@@ -196,6 +127,6 @@ class Graph(object):
             func: Function for metric to execute
         """
 
-        self.metrics.update({title: self.Metric(self.x_axis_label, self.y_axis_label, title, next(self.get_available_color), func)})
+        self.metrics.update({title: Metric(self.x_axis_label, self.y_axis_label, title, next(self.get_available_color), func)})
